Bug 310949 - Should display script URI with proper protocol

+++ bundles/org.eclipse.wst.jsdt.debug.core/src/org/eclipse/wst/jsdt/debug/internal/core/launching/JavaScriptSourceLookupParticipant.java
@@ -13,11 +13,12 @@ package org.eclipse.wst.jsdt.debug.internal.core.launching;
 import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
+import java.net.URI;
 
 import org.eclipse.core.resources.IFile;
+import org.eclipse.core.resources.IWorkspaceRoot;
 import org.eclipse.core.resources.ResourcesPlugin;
 import org.eclipse.core.runtime.CoreException;
-import org.eclipse.core.runtime.IPath;
 import org.eclipse.core.runtime.Path;
 import org.eclipse.core.runtime.URIUtil;
 import org.eclipse.debug.core.sourcelookup.AbstractSourceLookupParticipant;
@@ -55,20 +56,25 @@ public class JavaScriptSourceLookupParticipant extends AbstractSourceLookupParti
 	 * @see org.eclipse.debug.core.sourcelookup.AbstractSourceLookupParticipant#findSourceElements(java.lang.Object)
 	 */
 	public Object[] findSourceElements(Object object) throws CoreException {
-		String path = getSourcePath(object);
-		if (path != null) {
-			// TODO not sure if we should do a search for the member if the URI path is a miss
-			IFile file = (IFile) ResourcesPlugin.getWorkspace().getRoot().findMember(new Path(path), false);
-			if (file != null) {
-				return new IFile[] { file };
+		URI sourceURI = getSourceURI(object);
+		if (sourceURI != null) {
+			IWorkspaceRoot workspaceRoot = ResourcesPlugin.getWorkspace().getRoot();
+			URI workspaceURI = workspaceRoot.getRawLocationURI();			
+			URI workspaceRelativeURI = workspaceURI.relativize(sourceURI);
+			if (! workspaceRelativeURI.isAbsolute()) {
+				IFile file = (IFile) ResourcesPlugin.getWorkspace().getRoot().findMember(new Path(workspaceRelativeURI.getPath()), false);
+				if (file != null) {
+					return new IFile[] { file };
+				}
 			}
+
 			//try to find it using the source tab infos
 			Object[] sources = super.findSourceElements(object);
 			if(sources != null && sources.length > 0) {
 				return sources;
 			}
 			//else show the temp source
-			return showExternalSource(getSource(object), path);
+			return showExternalSource(getSource(object), sourceURI);
 		}
 		return NO_SOURCE;
 	}
@@ -78,7 +84,7 @@ public class JavaScriptSourceLookupParticipant extends AbstractSourceLookupParti
 	 * @param object
 	 * @return the raw source or <code>null</code>
 	 */
-	String getSource(Object object) {
+	private String getSource(Object object) {
 		if(object instanceof IJavaScriptStackFrame) {
 			IJavaScriptStackFrame jframe = (IJavaScriptStackFrame) object;
 			return jframe.getSource();
@@ -91,19 +97,19 @@ public class JavaScriptSourceLookupParticipant extends AbstractSourceLookupParti
 	}
 	
 	/**
-	 * Returns the path to use to look up source
+	 * Returns the uri to use to look up source
 	 * @param object
-	 * @return the path or <code>null</code>
+	 * @return the URI or <code>null</code>
 	 * @since 1.1
 	 */
-	String getSourcePath(Object object) {
+	private URI getSourceURI(Object object) {
 		if(object instanceof IJavaScriptStackFrame) {
 			IJavaScriptStackFrame jframe = (IJavaScriptStackFrame) object;
-			return jframe.getSourcePath();
+			return URI.create(jframe.getSourcePath());
 		}
 		if(object instanceof IScript) {
 			IScript script = (IScript) object;
-			return script.sourceURI().getPath().toString();
+			return script.sourceURI();
 		}
 		return null;
 	}
@@ -115,42 +121,37 @@ public class JavaScriptSourceLookupParticipant extends AbstractSourceLookupParti
 	 * @param path
 	 * @return the collection of files to show in external editors
 	 */
-	Object[] showExternalSource(String source, String path) {
+	private Object[] showExternalSource(String source, URI uri) {
 		try {
 			File tempdir = new File(System.getProperty("java.io.tmpdir")); //$NON-NLS-1$
-			File file = new File(tempdir, formatExternalName(path));
+			File jsdt_debug = new File(tempdir, "jsdt_debug"); //$NON-NLS-1$
+			jsdt_debug.mkdir();
+			jsdt_debug.deleteOnExit();
+
+			String fileName = URIUtil.lastSegment(uri);
+			if (fileName.endsWith(".js")) //$NON-NLS-1$
+				fileName = fileName.substring(0, fileName.length()-3);
+			fileName +="(" + Integer.toString(uri.toString().hashCode() + source.hashCode()) + ")"; //$NON-NLS-1$ //$NON-NLS-2$
+			fileName += ".js"; //$NON-NLS-1$
+			
+			File file = new File(jsdt_debug, fileName);
 			file.deleteOnExit();
-			FileWriter writer = new FileWriter(file);
-			writer.write(source);
-			writer.close();
+			
+			if (!file.exists()) {
+				FileWriter writer = new FileWriter(file);
+				writer.write(source);
+				writer.close();				
+			}
+			
+			if (JavaScriptDebugPlugin.getExternalScriptPath(fileName) == null)
+				JavaScriptDebugPlugin.addExternalScriptPath(fileName, uri.toString());
+				
 			return new Object[] {file};
 
 		} catch (IOException e) {
 			JavaScriptDebugPlugin.log(e);
 			return NO_SOURCE;
 		}
-	}
-
-	/**
-	 * Formats the name of the external source to be:
-	 * <code>[name]_[counter].js</code>.<br><br>
-	 * If the given name is <code>null</code> the default
-	 * name of <code>script_[counter].js</code> will be used
-	 * @param basename
-	 * @return the formatted external source name
-	 */
-	String formatExternalName(String basename) {
-		IPath path = null;
-		if(basename == null) {
-			path = new Path("script"); //$NON-NLS-1$
-		}
-		else {
-			path = new Path(basename);
-		}
-		path = path.removeFileExtension();
-		StringBuffer buffer = new StringBuffer(path.toString());
-		buffer.append(".js"); //$NON-NLS-1$
-		return buffer.toString();
 	}
 	
 	/*
